Slower version of transit skimming script
Our modeling machine fails to run 15 processes simultaneously, since the writing of the large ROUTEO files causes writes to fail.  This version splits up the jobs by time period only (as opposed to by time period and skim set) and runs the 3 skim sets in serial, and seems to succeed on MTC servers.